for extendability: renamed module to armed_bandits, created Greedy superclass of EpsilonGreedy, and implemented option to weigh reward based on recency
