+++ axelrod/tests/strategies/test_armed_bandits.py
